requirements.txt
dj-database-url==0.5.0
Django==2.1.2
django-heroku==0.3.1
gunicorn==19.9.0
pathlib==1.0.1
psycopg2==2.7.5
pytz==2018.7
PyYAML==3.13
whitenoise==4.1
django-ckeditor==5.6.1
pillow==5.3.0